+++ service/src/main/java/ar/edu/itba/paw/service/SeriesServiceImpl.java
@@ -125,7 +125,7 @@ public class SeriesServiceImpl implements SeriesService {
     }
 
     @Override
-    public List<Series> getAllSeriesByGenre(int id){
+    public List<Series> getAllSeriesByGenre(long id){
         return seriesDao.getSeriesByGenre(id);
     }
